move things around & prevent circular import
